+++ src/mayim/__init__.py
@@ -1,9 +1,8 @@
+from .base import Executor, Hydrator
 from .decorator import hydrator, register, sql
-from .executor import Executor
-from .executor.mysql import MysqlExecutor
-from .executor.postgres import PostgresExecutor
-from .executor.sqlite import SQLiteExecutor
-from .hydrator import Hydrator
+from .impl.sql.mysql.executor import MysqlExecutor
+from .impl.sql.postgres.executor import PostgresExecutor
+from .impl.sql.sqlite.executor import SQLiteExecutor
 from .mayim import Mayim
 
 __all__ = (
