+++ src/mayim/base/query.py
@@ -1,0 +1,6 @@
+from abc import ABC
+
+
+class Query(ABC):
+    __slots__ = ()
+    name: str
