--- src/mayim/executor/__init__.py
@@ -1,3 +1,0 @@
-from .base import Executor, is_auto_exec
-
-__all__ = ("Executor", "is_auto_exec")
