--- src/mayim/hydrator/__init__.py
@@ -1,3 +1,0 @@
-from .base import Hydrator
-
-__all__ = ("Hydrator",)
