+++ src/mayim/impl/sql/postgres/interface.py
@@ -1,8 +1,8 @@
 from contextlib import asynccontextmanager
 from typing import AsyncIterator, Optional
 
+from mayim.base.interface import BaseInterface
 from mayim.exception import MayimError
-from mayim.interface.base import BaseInterface
 
 try:
     from psycopg import AsyncConnection
