+++ src/mayim/impl/sql/sqlite/query.py
@@ -1,13 +1,13 @@
 import re
 
 from mayim.exception import MayimError
-from mayim.query.sql import ParamType, SQLQuery
+from mayim.impl.sql.query import ParamType, SQLQuery
 
 
-class PostgresQuery(SQLQuery):
+class SQLiteQuery(SQLQuery):
     __slots__ = ("name", "text", "param_type")
-    PATTERN_POSITIONAL_PARAMETER = re.compile(r"%s")
-    PATTERN_KEYWORD_PARAMETER = re.compile(r"%\([a-z_][a-z0-9_]*\)")
+    PATTERN_POSITIONAL_PARAMETER = re.compile(r"\?")
+    PATTERN_KEYWORD_PARAMETER = re.compile(r"\:[a-z_][a-z0-9_]")
 
     def __init__(self, name: str, text: str) -> None:
         super().__init__(name, text)
@@ -26,15 +26,3 @@ class PostgresQuery(SQLQuery):
             self.param_type = ParamType.KEYWORD
         else:
             self.param_type = ParamType.NONE
-
-    def convert_sql_params(self) -> str:
-        converted_text = self.text.replace("%", "%%")
-        if self.param_type == ParamType.POSITIONAL:
-            return self.PATTERN_POSITIONAL_PARAMETER.sub(
-                r"%s", converted_text, 0
-            )
-        if self.param_type == ParamType.KEYWORD:
-            return self.PATTERN_KEYWORD_PARAMETER.sub(
-                r"%(\2)s", converted_text, 0
-            )
-        return converted_text
