--- src/mayim/query/__init__.py
@@ -1,3 +1,0 @@
-from .base import Query
-
-__all__ = ("Query", "ParamType")
