+++ tests/test_init.py
@@ -1,9 +1,9 @@
 from unittest.mock import AsyncMock
 
 from mayim import Mayim, register
-from mayim.executor.base import Executor
-from mayim.interface.mysql import MysqlPool
-from mayim.interface.postgres import PostgresPool
+from mayim.base.executor import Executor
+from mayim.impl.sql.mysql.interface import MysqlPool
+from mayim.impl.sql.postgres.interface import PostgresPool
 from mayim.registry import Registry
